python/requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
build==1.0.3
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cookiecutter==2.5.0
crashtest==0.4.1
cryptography==42.0.2
descope==1.6.2
distlib==0.3.8
Django==5.0.2
django_descope==1.3.2
dnspython==2.5.0
dulwich==0.21.7
email-validator==2.1.0.post1
fastjsonschema==2.19.1
filelock==3.13.1
idna==3.6
importlib-metadata==7.0.1
installer==0.7.0
jaraco.classes==3.3.0
Jinja2==3.1.3
keyring==24.3.0
liccheck==0.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.7
numpy==1.26.4
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==4.2.0
poetry==1.7.1
poetry-core==1.9.0
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.3
pywatchman==1.4.1
PyYAML==6.0.1
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
scipy==1.12.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
summa==1.2.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.12.3
trove-classifiers==2024.1.31
types-python-dateutil==2.8.19.20240106
urllib3==2.2.0
virtualenv==20.25.0
xattr==1.1.0
zipp==3.17.0
